requirements.txt
flask
flask_sqlalchemy
argon2-cffi
mysqlclient
uvicorn
gunicorn
gevent
asgiref
python-dotenv
